requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==2.2.28
django-babel==0.6.2
pytz==2025.1
setuptools==75.8.0
sqlparse==0.5.3
wheel==0.45.1
zxcvbn-python==4.4.24
